extract section view model
